requirements.txt
Django 
djangorestframework==3.12.4
mysql_connector_repackaged==0.3.1
pytz==2021.1
django-crispy-forms==2.1
django-filter==23.5
django-restapi==0.1.3
django-cors-headers==3.7.0
django-crispy-forms==2.1

 